# Optimize elementwise_add_grad op