<!-- Demo: https://github.com/PaddlePaddle/Paddle/pull/24810 -->
### PR types
<!-- One of [ New features | Bug fixes | Function optimization | Performance optimization | Breaking changes | Others ] -->
Performance optimization
### PR changes
<!-- One of [ OPs | APIs | Docs | Others ] -->
OPs
### Describe
<!-- Describe what this PR does -->
# 起因
`SimpleElemwiseAddGradCUDAKernel`的实现为拷贝`dout`的值到`dx`和`dy`中。然而，此处存在一处可优化点：即若`dout`与`dx`或`dy`复用了同一片地址，则无需多余拷贝。

# 优化方案
1. 当`dx_data`与`dout_data`相同且当`dy_data`与`dout_data`不同时：只需要拷贝`dout_data`到`dy_data`即可。
2. 当`dx_data`与`dout_data`不同且当`dy_data`与`dout_data`相同时：只需要拷贝`dout_data`到`dx_data`即可。
3. 当`dx_data`和`dy_data`与`dout_data`都不同时：调用原`SimpleElemwiseAddGradCUDAKernel`。
4. 当`dx_data`和`dy_data`与`dout_data`都相同时：什么都不用做。

> 问：为什么放在`elementwise_add_op.cu`这儿而不是放在`ElementwiseAddGradKernel::Compute`处？
> 答：因为会报错`Tensor not alloc memory`

# 优化结果
`elementwise_add_grad` op耗时 | 优化前 | 优化后
---|---|---|
float16 | 147.551 us | 80.672 us
float | 244.958 us | 159.391 us